requirements.txt
matplotlib
scikit-image
scikit-learn
pandas
numpy
spotipy
pyyaml
tqdm
autopep8
fastapi
uvicorn
pytest
httpx
pytest-mock 
pytest-asyncio